Create Car.java
Coursera
Amazon
Activity: Using a linear search
Have you ever wondered how online marketplaces allow you to find the item you want among thousands of products, all in a matter of seconds? One of the simplest methods to do this is a linear search. Remember, a linear search traverses a collection of values or objects sequentially, comparing each to some criteria. It is the slowest search algorithm but the simplest to implement and is particularly useful for a small collection of unordered objects. In this lab, you will implement a linear search on an array. Let’s get started!

Imagine you're a developer at BuyNow Automotive Dealership. The dealership requires a reliable system to search its inventory of cars for potential buyers. They want to be able to search for cars by make and model, color, availability, and price range. As the resident software developer, your task is to implement the search functionalities to help sales representatives quickly find the right cars for their customers. 

Goal

The CarSearch application is used by the BuyNow Automotive Dealership to find cars for sale by make and model, color, availability, or price range. You will implement three search operations and search for cars by: 

🖥️  Note: When you encounter this icon, it’s time to get into your IDE and start coding!

In your lab environment, open IntelliJ by double-clicking on the icon.

When you open IntelliJ, the starter code will be displayed. You should see two classes, Car and CarSearch. The domain object used in this lab is the Car. It represents the make, model, color, availability, and price of a car sold by BuyNow Automotive Dealership.