Create Database.java
Coursera
Amazon
Activity: Using Bubble Sort
Introduction

In previous labs, you helped the State of New York build an automobile registration system by implementing a binary search to locate license numbers. The system worked well, thanks to a pre-sorted array of Registration objects provided by the Database class. Now, they’ve asked for further improvements because the system isn't as efficient as they’d like.

To boost performance, you've decided to implement a Bubble Sort algorithm because it's effective, easy to implement, and because of its simplicity it is easy to maintain compared to other sorting algorithms. A Bubble Sort repeatedly steps through a list, compares adjacent elements, and swaps them if needed, with smaller elements 'bubbling' to the top and larger ones sinking to the bottom until the list is fully sorted.

Goal

The RegistrationSearch application is used by the State of New York Department of Motor Vehicles (DMV) to access the registration information associated with license plate numbers. Critical to the application's success is the sorting of registration information by license plate numbers. As the software developer in charge, it's your job to implement a Bubble Sort to ensure the data is sorted properly. You will implement a single search operation that takes a license number as a parameter and returns a registration object associated with that license number.