Add Main class with main method
Amazon
Programming Assignment: Graded Lab: Basic testing
You have not submitted. You must earn 80/100 points to pass.
Deadline
Pass this assignment by Dec 15, 11:59 PM IST

In this graded lab, you’ll take on the exciting challenge to build a contact book application using TDD. This is your chance to put your new skills into practice! A contact book is a tool used to manage personal or professional connections, so your task is to ensure it functions correctly under various conditions. 

Throughout this lab, you’ll write JUnit test cases to validate the core features of the contact book, such as adding, removing, and searching for contacts. Each task will guide you through different phases of TDD, ensuring that every feature is thoroughly tested and reliable.

Goal

Your mission is to develop the contact book using the TDD approach, following the Red-Green-Refactor cycle. You will start by writing tests for functionalities that do not yet exist (Red Phase), then implement the code to make the tests pass (Green Phase), and finally, refine and optimize the code (Refactor Phase).

Requirements 
In this lab, you will: 

Create a Contact Book that can store the names and phone numbers of contacts. It is represented by the class ContactBook.

Use arrays to store the Contact Book, meaning there is a limit to the maximum number of contacts that can be stored. 

Represent each contact by a class Contact. 

Add test methods in the ContactBookTest class to ensure validation of requirements to add, remove, and search a contact.

Each of these methods shall be added using the Red-Green-Refactor cycle of TDD. This means that if the requirements are falling short, you will modify the ContactBook class to meet the requirements. 

🖥️  Note: When you encounter this icon, it’s time to get into your IDE and start coding!

In your lab environment, open IntelliJ by double-clicking on the icon.


When you open the starter code, apart from the other folders, you will find three folders: 

src: This folder contains the source code for your program consisting of the following files:

Main.java, which is empty since you are focusing on the tests first.

Contact.java, which contains the class Contact, consisting of the properties Name and PhoneNumber. 

ContactBook.java, which contains the class ContactBook. The class contains the list of contacts. 

test: This folder holds the test class named ContactBookTest inside the file ContactBookTest.java. 

Lib: This folder contains the JUnit 5 test library jar files along with the required dependencies inside it. You do not need to do anything with this folder.

Now, let’s explore the following classes in more detail:

Contact 

ContactBook

ContactBookTest 

The Contact class
The class named Contact in the file Contact.java contains the representation of a contact, consisting of the properties name and phoneNumber. It also includes a default constructor (which is a constructor with arguments)  and overloaded constructor (which is a constructor with arguments) , and getters and setters (methods to get the value of the properties and change the value of properties of the class) for the properties. 

The ContactBook class
The class named ContactBook in the file ContactBook.java contains the basic code setup of the class containing the contacts. A final static variable named MAX_CONTACTS of type int sets the maximum number of contacts that can be stored by the contact book, which is 5. 

Note: The contacts array is allocated a maximum memory of 5 contacts in the default constructor. 

The variable named numberOfContacts keeps track of the contacts currently being held in the contact book. There are no other methods in the class ContactBook. They will be implemented during the TDD cycle. 

The ContactBookTest class
The test class ContactBookTest in the file ContactBookTest.java contains the basic code setup and empty test methods for TDD testing. You will be instructed whether to work in the file ContactBook.java or ContactBookTest.java. The Contact class is completed and does not require action.

Please take note of the following:

Every task in this Graded Lab will have a subheading titled "Phase". This indicates whether it is a Red Phase, a Green Phase, or a Refactor Phase, which will make it easier for you to understand.

Ignore the warnings you get when you open the starter code. All references to the method addContact() will appear red in ContactBookTest.java. This is because the method does not exist. This error will go away as soon as you complete TODO 1.

You will run each test method individually, as this is how they have been set up.